refactor: recoil-persist를 사용하여 로컬스토리지에 atom 동기화
