+++ package.json
@@ -9,6 +9,7 @@
     "react-router-dom": "^6.11.1",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
+    "recoil-persist": "^4.2.0",
     "styled-components": "^6.0.0-rc.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
