+++ src/constants/index.ts
@@ -2,3 +2,7 @@ export const PRODUCT = {
   MIN_COUNT: 0,
   MAX_COUNT: 99,
 } as const;
+
+export const LOCAL_STORAGE_KEY = {
+  CART: 'cart',
+} as const;
